parser/model: move reorg related types to a new file
